Authentication and Authorization added with accessToken and refreshToken make some routes public and put some routes under private all security stuffs in Security packege and create AuthController inside controller packege with signup , login (and refreshToken end pont also) this is a simple name email and password based signup only by default every user who is signup became the GuestRole

+++ src/main/java/com/soumyajit/HotelBooking/service/HotelServiceImpl.java
@@ -1,16 +1,19 @@
 package com.soumyajit.HotelBooking.service;
 
 import com.soumyajit.HotelBooking.Exception.ResourceNotFound;
+import com.soumyajit.HotelBooking.Exception.UnAuthorizedException;
 import com.soumyajit.HotelBooking.dtos.HotelDTOS;
 import com.soumyajit.HotelBooking.dtos.HotelInfoDTOS;
 import com.soumyajit.HotelBooking.dtos.RoomDTOS;
 import com.soumyajit.HotelBooking.entities.Hotel;
 import com.soumyajit.HotelBooking.entities.Room;
+import com.soumyajit.HotelBooking.entities.User;
 import com.soumyajit.HotelBooking.repository.HotelRepository;
 import com.soumyajit.HotelBooking.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +34,10 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating new Hotel with name:{}" ,hotelDTOS.getName());
         Hotel hotel = modelMapper.map(hotelDTOS,Hotel.class);
         hotel.setIsActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         Hotel savedHotel  = hotelRepository.save(hotel);
         log.info("Created a new Hotel with this : {}",hotelDTOS.getId());
         return modelMapper.map(savedHotel,HotelDTOS.class);
@@ -42,6 +49,13 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFound("Hotel not found with ID :"+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())){     //if the Admin isn't won the hotel if we don't put this any admin can access this
+            throw new UnAuthorizedException("This user doesn't won this hotel with id: "+id);
+        }
+
         return modelMapper.map(hotel,HotelDTOS.class);
     }
 
@@ -61,7 +75,14 @@ public class HotelServiceImpl implements HotelService {
         if(!isExists(id)){
             throw new ResourceNotFound("Hotel not found with ID :"+id);
         }
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         Hotel hotel = modelMapper.map(hotelDTOS, Hotel.class);
+
+        if(!user.equals(hotel.getOwner())){     //if the Admin isn't won the hotel if we don't put this any admin can access this and can update this
+            throw new UnAuthorizedException("This user doesn't won this hotel with id: "+id);
+        }
         hotel.setId(id);
         Hotel savedHotel = hotelRepository.save(hotel);
         return modelMapper.map(savedHotel, HotelDTOS.class);
@@ -73,6 +94,12 @@ public class HotelServiceImpl implements HotelService {
         log.info("Deleting Hotel with this : {}",hotelId);
         Hotel hotel = hotelRepository.findById(hotelId)
                 .orElseThrow(()->new ResourceNotFound("Hotel not found with ID :"+hotelId));
+
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){     //if the Admin isn't won the hotel, if we don't put this any admin can access this and can delete this
+            throw new UnAuthorizedException("This user doesn't won this hotel with id: "+hotelId);
+        }
 
         for(Room room : hotel.getRoom()){
             inventoryService.deleteAllInventories(room);
@@ -90,6 +117,11 @@ public class HotelServiceImpl implements HotelService {
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFound("Hotel not found with ID: "+hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){     //if the Admin isn't won the hotel, if we don't put this any admin can access this and can activate this
+            throw new UnAuthorizedException("This user doesn't won this hotel with id: "+hotelId);
+        }
+
         hotel.setIsActive(true);
 
         if(hotel.getIsActive()){
@@ -102,11 +134,12 @@ public class HotelServiceImpl implements HotelService {
     }
 
     @Override
-    public HotelInfoDTOS getHotelInfoById(Long hotelId) {
+    public HotelInfoDTOS getHotelInfoById(Long hotelId) { //public route
         log.info("Getting the hotel information with ID: {}", hotelId);
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFound("Hotel not found with ID: "+hotelId));
+
         List<RoomDTOS> roomDTOS = hotel.getRoom().stream().map(room ->
                 modelMapper.map(room, RoomDTOS.class))
                 .toList();
